# perf: Canvas rendering optimization (#61)